+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-11T13:44:33.441Z",
+    "2026-05-11T15:08:58.855Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {