+++ packages/orq-rc/RELEASES.md
@@ -9039,3 +9039,13 @@ Based on:
 - [typescript v4.9.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.42 - packages/orq-rc
+
+## 2026-05-11 15:12:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.43 - packages/orq-rc